Moved json-related function to a separated file
*code refactored
*added new required field to the json schema (id)
